Add files via upload
Backtracking is finding a solution to a problem where the solution depends on the steps taken previously. For example, in the maze problem, the solution depends on all the steps you take one by one. If any of those steps are wrong, then it will not lead us to a solution. In the maze problem, we first choose a path and keep moving along it. But once we understand that a certain path is wrong, then we just go back and change it. This is what is basically backwards.

In backtracking, we first take a step and then see whether this step taken is correct or not, i.e. whether to give the correct answer or not. And if not, then we go back and change our first move. In general, this is achieved by recursion. So, in backwards, we first start with a partial sub-solution of the problem (which may not lead us to a solution) and then whether we can proceed further with this sub-solution or not. If not, then we just go back and change it.